Fix/update machine fix

+++ src/api/theoreticalMachine/theoreticalMachineService.ts
@@ -61,6 +61,10 @@ export const theoreticalMachineService = {
         return new ServiceResponse(ResponseStatus.Failed, messages.userNotFound, null, StatusCodes.NOT_FOUND);
       }
       const machineIdAsNumber = parseInt(machineId, 10);
+      const machineExists = await theoreticalMachineRepository.getMachineByIdAsync(userId, machineIdAsNumber);
+      if (!machineExists) {
+        return new ServiceResponse(ResponseStatus.Failed, messages.machineNotFound, null, StatusCodes.NOT_FOUND);
+      }
       const isSuccessful = await theoreticalMachineRepository.deleteMachineAsync(userId, machineIdAsNumber);
       if (!isSuccessful) {
         throw new Error(messages.machineDeleteFailed);
@@ -83,6 +87,10 @@ export const theoreticalMachineService = {
         return new ServiceResponse(ResponseStatus.Failed, messages.userNotFound, null, StatusCodes.NOT_FOUND);
       }
       const machineIdAsNumber = parseInt(machineId, 10);
+      const machineExists = await theoreticalMachineRepository.getMachineByIdAsync(userId, machineIdAsNumber);
+      if (!machineExists) {
+        return new ServiceResponse(ResponseStatus.Failed, messages.machineNotFound, null, StatusCodes.NOT_FOUND);
+      }
       const isSuccessful = await theoreticalMachineRepository.updateUserMachineAsync(
         userId,
         machineIdAsNumber,
